fixed Encoder/Decoder to do the property type conversion for unsigned types
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk@559520 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/java/common/src/main/java/org/apache/qpidity/Decoder.java
@@ -33,9 +33,9 @@ public interface Decoder
 {
 
     boolean readBit();
-    byte readOctet();
-    short readShort();
-    int readLong();
+    short readOctet();
+    int readShort();
+    long readLong();
     long readLonglong();
 
     long readTimestamp();
@@ -44,7 +44,7 @@ public interface Decoder
     String readLongstr();
 
     Map<String,?> readTable();
-    Range<Integer>[] readRfc1982LongSet();
+    Range<Long>[] readRfc1982LongSet();
     UUID readUuid();
 
     String readContent();
